refactor: move theme boot from index.html to main entry
Extract resolveTheme/applyDocumentTheme into lib/theme.ts and run
apply-initial-theme as the first main.tsx import before styles.css.
