+++ frontend/index.html
@@ -4,22 +4,6 @@
 		<meta charset="UTF-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<title>Agent Orchestrator</title>
-		<script>
-			(function () {
-				var stored = null;
-				try {
-					stored = localStorage.getItem("ao.theme");
-				} catch (e) {}
-				var theme =
-					stored === "light" || stored === "dark"
-						? stored
-						: window.matchMedia && window.matchMedia("(prefers-color-scheme: light)").matches
-							? "light"
-							: "dark";
-				document.documentElement.dataset.theme = theme;
-				document.documentElement.style.colorScheme = theme;
-			})();
-		</script>
 	</head>
 	<body>
 		<div id="root"></div>
