+++ frontend/src/renderer/main.tsx
@@ -1,3 +1,4 @@
+import "./lib/apply-initial-theme";
 import React from "react";
 import { createRoot } from "react-dom/client";
 import { QueryClientProvider } from "@tanstack/react-query";
